requirements.txt
Flask==2.2.2
requests==2.28.2
Werkzeug==2.2.3
flask-cors==3.1.1 
python-dotenv==0.21.0 
